Move action utilities to their own module
